+++ plugin/src/net/groboclown/idea/p4ic/config/part/FileDataPart.java
@@ -165,6 +165,7 @@ public class FileDataPart implements DataPart {
         validation.checkPort(this, rawPort);
         validation.checkAuthTicketFile(this);
         validation.checkTrustTicketFile(this);
+        validation.checkClientName(this, false);
         return validation.getProblems();
     }