+++ plugin/src/net/groboclown/idea/p4ic/ui/config/ResolvedPropertiesPanel.java
@@ -233,13 +233,6 @@ public class ResolvedPropertiesPanel {
                 new BackgroundAwtActionRunner.BackgroundAwtAction<String>() {
                     @Override
                     public String runBackgroundProcess() {
-                        // FIXME reloading the config is important, but refreshing just the
-                        // one config isn't the right way to do this.  We need to refresh
-                        // the whole stack of configs.  Note that it might change the
-                        // paths, though.
-                        // if (selected.config != null) {
-                        //     selected.config.reload();
-                        // }
                         StringBuilder sb = createResolvedPropertiesText(selected);
                         return sb.toString();
                     }